requirements.txt
streamlit>=1.28.0
neo4j>=5.0.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
peft>=0.4.0
accelerate>=0.20.0
datasets>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0 
# LangChain 패키지들
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.28
langchain-experimental>=0.0.49
# 한국어 형태소 분석
konlpy>=0.6.0
# 추가 유틸리티
tqdm>=4.64.0
selenium>=4.0.0
selenium-stealth>=1.0.6 